creating profile handler

+++ interface/remote_actions.py
@@ -23,7 +23,10 @@ class S3Interface:
         remote_files = self.get_remote_file_names(bucket_name)
         local_files = directory_files_recursively(local_path)
         files_to_delete = get_files_to_delete(local_files, remote_files)
-        return self.delete_aws_files(files_to_delete, bucket_name)
+        if files_to_delete.get('Objects')
+            return self.delete_aws_files(files_to_delete, bucket_name)
+        else:
+            return []
 
     def get_remote_files(self, bucket_name: str) -> list:
         """
